+++ js/src/lib/query-reader.ts
@@ -40,7 +40,6 @@ export class ConsoleQueryReader implements IQueryReader {
       let scriptName = "query" + i;
       let match = script.match(/^([\d\w]+)\:/);
       if (match && match.length > 1) {
-        console.log(match);
         scriptName = match[1];
         script = script.substring(scriptName.length + 1);
       }